requirements.txt
python-dateutil>=2,<3
toolz>=0.10,<1
xmltodict>=0.12,<1.0
requests>=2,<3
typing-extensions>=3,<4